Comprobaciones
Esta comprobado: nomal, 1a forma, 2a forma, curvaturas: principales, media y gauss
